+++ src/shared/components/input/input.type.ts
@@ -1,0 +1,14 @@
+import type { InputHTMLAttributes, ReactNode } from 'react';
+
+interface IconProps {
+  icon: ReactNode;
+  onIconClick?: () => void;
+}
+
+interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
+  label: string;
+  id: string;
+  iconProps?: IconProps;
+}
+
+export type { InputProps, IconProps };
